imdbapi/requirement.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11.15
djangorestframework==3.8.2
idna==2.7
lxml==4.2.4
omdb==0.10.0
pkg-resources==0.0.0
pytz==2018.5
requests==2.19.1
SQLAlchemy==1.2.11
urllib3==1.23
